services/fyers-bridge/requirements.txt
fyers-apiv3
fastapi
uvicorn[standard]
python-dotenv
